restructure lib components and update ts config
