+++ app/components/ResumeSection.tsx
@@ -1,16 +1,16 @@
 import Link from 'next/link'
-import ThemeImage from './ThemeImage'
-import tpcLight from '../../public/tpc_light.png'
-import tpcDark from '../../public/tpc_dark.png'
-import starbucksLight from '../../public/starbucks_light.png'
-import starbucksDark from '../../public/starbucks_dark.png'
-import microsoftLight from '../../public/microsoft_light.png'
-import microsoftDark from '../../public/microsoft_dark.png'
-import boeingLight from '../../public/boeing_light.png'
-import boeingDark from '../../public/boeing_dark.png'
-import highSeasLight from '../../public/high-seas_light.png'
-import highSeasDark from '../../public/high-seas_dark.png'
-import Button from './Button'
+import ThemeImage from '@/lib/ThemeImage'
+import tpcLight from '@public/tpc_light.png'
+import tpcDark from '@public/tpc_dark.png'
+import starbucksLight from '@public/starbucks_light.png'
+import starbucksDark from '@public/starbucks_dark.png'
+import microsoftLight from '@public/microsoft_light.png'
+import microsoftDark from '@public/microsoft_dark.png'
+import boeingLight from '@public/boeing_light.png'
+import boeingDark from '@public/boeing_dark.png'
+import highSeasLight from '@public/high-seas_light.png'
+import highSeasDark from '@public/high-seas_dark.png'
+import Button from '@/lib/Button'
 
 export default function ResumeSection() {
 	return (
