+++ app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next'
 import './globals.css'
-import Header from './components/Header'
-import Providers from './components/Providers'
-import Footer from './components/Footer'
+import Header from '@/components/Header'
+import Providers from '@/components/Providers'
+import Footer from '@/components/Footer'
 import { Analytics } from '@vercel/analytics/react'
 import { SpeedInsights } from '@vercel/speed-insights/next'
