+++ app/projects/[handle]/page.tsx
@@ -1,4 +1,4 @@
-export default function About() {
+export default function ProjectDetail() {
 	return (
 		<div className="w-full h-full flex flex-col items-center justify-center text-center p-8">
 			<h1 className="font-sans text-5xl font-medium mb-2">ABOUT</h1>
