--- app/projects/page.tsx
@@ -1,7 +1,0 @@
-export default function About() {
-	return (
-		<div className="w-full h-full flex flex-col items-center justify-center text-center p-8">
-			<h1 className="font-sans text-5xl font-medium mb-2">ABOUT</h1>
-		</div>
-	)
-}
